backend/requirements.txt
falcon==3.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.3
gunicorn==21.2.0
falcon-marshmallow==0.4.0
marshmallow==3.20.1
python-dateutil==2.8.2 